Add superworking recursive reflections

+++ include/Camera.h
@@ -36,7 +36,7 @@ private:
     //2D 1000x1000 array of type Pixel
     std::array<std::array<Pixel, HEIGHT>, WIDTH> pixels;
     // FOV in radians
-    float fov = M_PI / 2.0f;
+    float fov = ((float) M_PI) / 2.0f;
     //A bool to flip between the cameras
     bool eyeBeingUsed;
 };
